requirements.txt
Django==1.5.4
Python-Travis-Client==1.0
argparse==1.2.1
distribute==0.6.24
httplib2==0.8
keyring==3.0.5
launchpadlib==1.10.2
lazr.authentication==0.1.2
lazr.restfulclient==0.13.3
lazr.uri==1.0.3
oauth==1.0.1
psycopg2==2.5.1
python-dateutil==2.1
simplejson==3.3.0
six==1.4.1
testresources==0.2.7
wadllib==1.3.2
wsgi-intercept==0.5.1
wsgiref==0.1.2
zope.interface==4.0.5
